refactor : structure of Header
